Use Bcast instead of Allgatherv in assembleDistZVec()
- temporary fix: clone communicator inside permutations to avoid
  scheduling issues of the collective operation AlltoAll that is used
  internally
- disentagle the circular dependency between merge.h and rot.h by
  refactoring GivensRotation<T> and TileCollector
- only dlaf.core is needed to build and link `test_tridiag_solver_rot`
