+++ test/unit/eigensolver/CMakeLists.txt
@@ -78,7 +78,7 @@ DLAF_addTest(
 DLAF_addTest(
   test_tridiag_solver_rot
   SOURCES test_tridiag_solver_rot.cpp
-  LIBRARIES dlaf.core dlaf.tridiagonal_eigensolver dlaf.multiplication dlaf.permutations
+  LIBRARIES dlaf.core
   USE_MAIN MPIPIKA
   MPIRANKS 6
 )
